requirements/base.txt
Django==4.1.1
djangorestframework==3.13.1
celery==5.2.7
uuid==1.30

mutagen==1.45.1
pytube==12.1.0
ytmusicapi==0.22.0
pydub
Pillow
redis